refactor: Players 패키지 변경
