+++ README.md
@@ -80,5 +80,5 @@ example feat: 사용자 입력 후 도메인 사용
     - 딜러는 덱을 필드로 갖는다.
 - [x] 딜러의 이름 필드를 제거한다.
 - [ ] 공통된 코드 추가 메서드로 묶는다.
-- [ ] `protected` 필드 직접 접근 문제 해결한다.
+- [x] `protected` 필드 직접 접근 문제 해결한다.
 - [ ] `Participant`를 `Gamer`로 변경한다.
